Refactor some of the processing code which uses pythainlp
