+++ stanza/utils/datasets/tokenization/convert_th_best.py
@@ -28,7 +28,7 @@ import sys
 
 from pythainlp import sent_tokenize
 
-from stanza.utils.datasets.tokenization.process_thai_tokenization import write_dataset
+from stanza.utils.datasets.tokenization.process_thai_tokenization import reprocess_lines, write_dataset, convert_processed_lines
 
 def clean_line(line):
     line = line.replace("html>", "html|>")
@@ -76,44 +76,6 @@ def clean_word(word):
         return word
     return word
 
-def reprocess_lines(processed_lines):
-    reprocessed_lines = []
-    for line in processed_lines:
-        text = "".join(line)
-        chunks = sent_tokenize(text)
-        if sum(len(x) for x in chunks) != len(text):
-            raise ValueError("Got unexpected text length: \n{}\nvs\n{}".format(text, chunks))
-
-        chunk_lengths = [len(x) for x in chunks]
-
-        current_length = 0
-        new_line = []
-        for word in line:
-            if len(word) + current_length < chunk_lengths[0]:
-                new_line.append(word)
-                current_length = current_length + len(word)
-            elif len(word) + current_length == chunk_lengths[0]:
-                new_line.append(word)
-                reprocessed_lines.append(new_line)
-                new_line = []
-                chunk_lengths = chunk_lengths[1:]
-                current_length = 0
-            else:
-                remaining_len = chunk_lengths[0] - current_length
-                new_line.append(word[:remaining_len])
-                reprocessed_lines.append(new_line)
-                word = word[remaining_len:]
-                chunk_lengths = chunk_lengths[1:]
-                while len(word) > chunk_lengths[0]:
-                    new_line = [word[:chunk_lengths[0]]]
-                    reprocessed_lines.append(new_line)
-                    word = word[chunk_lengths[0]:]
-                    chunk_lengths = chunk_lengths[1:]
-                new_line = [word]
-                current_length = len(word)
-        reprocessed_lines.append(new_line)
-    return reprocessed_lines
-
 def read_data(input_dir):
     subdirs = [os.path.join(input_dir, 'article'),
                os.path.join(input_dir, 'encyclopedia'),
@@ -128,7 +90,6 @@ def read_data(input_dir):
     documents = []
     for filename in files:
         with open(filename) as fin:
-            sentences = []
             processed_lines = []
             for line in fin.readlines():
                 line = clean_line(line)
@@ -141,28 +102,9 @@ def read_data(input_dir):
                 processed_lines.append(words)
 
             processed_lines = reprocess_lines(processed_lines)
+            paragraphs = convert_processed_lines(processed_lines)
 
-            for words in processed_lines:
-                # turn the words into a sentence
-                sentence = []
-                for word in words:
-                    word = word.strip()
-                    if not word:
-                        if len(sentence) == 0:
-                            raise ValueError("Unexpected space at start of sentence in document {}".format(filename))
-                        sentence[-1] = (sentence[-1][0], True)
-                    else:
-                        sentence.append((word, False))
-                # blank lines are very rare in best, but why not treat them as a paragraph break
-                if len(sentence) == 0:
-                    paragraphs = [sentences]
-                    documents.append(paragraphs)
-                    sentences = []
-                    continue
-                sentence[-1] = (sentence[-1][0], True)
-                sentences.append(sentence)
-            paragraphs = [sentences]
-            documents.append(paragraphs)
+            documents.extend(paragraphs)
 
     return documents
